List Benchmark Support
This patch adds List Object (GET container)
benchmarking.
Users are able to define the ratio of List as
5th argument of crud profile.
A sample scenario for "list only benchmarking"
is as follows.

[Sample]
{
  "name": "List only test scenario",
  "sizes": [{
    "name": "tiny",
    "size_min": 4096,
    "size_max": 8192
  }],
  "initial_files": {
    "tiny": 1000
  },
  "operation_count": 500,
  "crud_profile": [0, 0, 0, 0, 1],
                              ^^^
  "user_count": 4,
  "container_base": "ssbench",
  "container_count": 10,
  "container_concurrency": 10
}

This patch keeps backward compatibility of scenario file,
so users are able to use old scenario files which have
no 5th argument for list at crud profile.